perf(updateRenderedImage): fuse images in web worker
